-fixed viewable_by update -added check in delete entry

+++ Registry/Registry.cpp
@@ -199,6 +199,8 @@ int Registry::view_imagelist_svc(std::vector<std::string> &image_container, long
 
     if (viewable_by_DB.empty())
         return -1;
+    else
+        update_viewable_by();
 
     auto n = check_token_svc(token);
     update_imageList();
@@ -247,7 +249,7 @@ int Registry::add_entry_svc(std::string image_name, long int token, sockaddr_in 
 
         }
         catch (std::exception &e) {
-            std::cout << "exception: " << e.what() << std::endl;
+            std::cout << "add_entry_svc exception: " << e.what() << std::endl;
         }
 
     } else {
@@ -260,9 +262,15 @@ int Registry::add_entry_svc(std::string image_name, long int token, sockaddr_in 
 int Registry::remove_entry_svc(std::string image_name, long int token) {
 
     auto n = check_token_svc(token);
+    bool check_owner = false;
+
+    for ( int i = 0; i < img_DB.size(); i++)
+        if (img_DB[i].img_name == image_name)
+            if (img_DB[i].token == token)
+                check_owner = true;
 
 //    const char const_char_image_name = image_name;
-    if (n == 0) {
+    if (n == 0 && check_owner) {
         try {
             SQLite::Database db(pathLocation, SQLite::OPEN_READWRITE, 0, "");
             SQLite::Statement view_query(db, "DELETE FROM viewable_by WHERE img_name ='" + image_name + "';");
@@ -270,10 +278,12 @@ int Registry::remove_entry_svc(std::string image_name, long int token) {
             SQLite::Statement img_query(db, "DELETE FROM image WHERE Img_name ='" + image_name + "' and token= " +
                                             std::to_string(token) + " ;");
             int noRowsModified = img_query.exec();
+            update_viewable_by();
+            update_imageList();
             return 0;
         }
         catch (std::exception &e) {
-            std::cout << "exception: " << e.what() << std::endl;
+            std::cout << "remove_entry_svc exception: " << e.what() << std::endl;
         }
 
     } else {
@@ -346,8 +356,8 @@ int Registry::check_viewImage_svc(std::string image_id, bool &can_view, long int
 int Registry::set_image_viewable_by_svc(std::string image_id, long int user_token, std::string allowed_user) {
     update_imageList();
     long int peer_token = fetch_token(allowed_user);
-    if (peer_token < 0)
-        return -1;
+//    if (peer_token < 0)
+//        return -1;
     for (int i = 0; i < img_DB.size(); i++)
         if (img_DB[i].token == user_token && img_DB[i].img_name == image_id) {
             try {
@@ -358,7 +368,7 @@ int Registry::set_image_viewable_by_svc(std::string image_id, long int user_toke
                 return 0;
             }
             catch (std::exception &e) {
-                std::cout << "exception: " << e.what() << std::endl;
+                std::cout << "set_image_viewable_by_svc exception: " << e.what() << std::endl;
             }
         }
 
@@ -414,7 +424,7 @@ void Registry::load_DBs() {
 
     }
     catch (std::exception &e) {
-        std::cout << "exception: " << e.what() << std::endl;
+        std::cout << "load_DBs exception: " << e.what() << std::endl;
     }
 
 }
@@ -453,7 +463,7 @@ void Registry::update_users() {
         }
     }
     catch (std::exception &e) {
-        std::cout << "exception: " << e.what() << std::endl;
+        std::cout << " update_users exception: " << e.what() << std::endl;
     }
